refactor(agent-runtime): split A2aJsonRpcHandler into request/response mappers
Extract inbound JSON-RPC params parsing into A2aRequestMapper and outbound A2A
wire normalization into A2aResponseMapper; A2aJsonRpcHandler retains only the
JSON-RPC envelope dispatch. Verbatim, behavior-preserving method moves — the
520-line file becomes three cohesive classes.

Verified on WSL: clean compile + checkstyle; non-Mockito a2a tests green;
deterministic A2A ping/pong e2e green. The Mockito-based A2aJsonRpcHandlerTest
validates in CI (this WSL cannot self-attach the ByteBuddy agent).
